remove skill logic from mission executor
