+++ src/skills/document/voucher_lines_parsing.ts
@@ -3,6 +3,8 @@ import { IPseudoTask, IPseudoMission } from "@/skills/types";
 import { ChatService } from "@/services/chat.service";
 import { prepareDocumentContext } from "@/skills/utils/document_helper";
 import { SchemaType, Schema } from "@google/generative-ai";
+import { MoneyUtil } from "@/lib/utils/money";
+import { VendorRegistry } from "@/services/rules/vendor_registry";
 
 export class VoucherLinesParsingSkill implements ITaskSkill {
   name = "VOUCHER_LINES_PARSING";
@@ -28,6 +30,7 @@ export class VoucherLinesParsingSkill implements ITaskSkill {
     mission: IPseudoMission,
     fullPrompt: string,
     chatService: ChatService,
+    priorResults?: Map<string, string>,
   ): Promise<string> {
     const { images, parsedContext } = await prepareDocumentContext(task);
 
@@ -36,6 +39,58 @@ export class VoucherLinesParsingSkill implements ITaskSkill {
 
     if (parsedContext.journalText) {
       promptText += `\n\n【重要指示】\n使用者已提供/修正日記帳的最新內容如下。請優先依據以下文字資訊進行解析，若與圖片內容有衝突，以此文字為準：\n${parsedContext.journalText}`;
+    }
+
+    // Info: (20260511 - Tzuhan) Stage 2 Deterministic Routing Intercept
+    if (priorResults) {
+      let baseParsed: Record<string, unknown> | null = null;
+      for (const prevResultStr of priorResults.values()) {
+        try {
+          const parsed = JSON.parse(prevResultStr);
+          const actualParsed = parsed.data || parsed;
+
+          if (actualParsed.error) {
+            throw new Error(`AI 解析失敗: ${actualParsed.error}`);
+          }
+
+          if (actualParsed.vendorName && actualParsed.documentType) {
+            baseParsed = actualParsed;
+            break;
+          }
+        } catch (err) {
+          if (err instanceof Error && err.message.includes("AI 解析失敗")) {
+            throw err;
+          }
+        }
+      }
+
+      if (baseParsed && baseParsed.vendorName) {
+        const matchedRules = VendorRegistry.match(
+          String(baseParsed.vendorName),
+          String(baseParsed.documentType || "BILL_NOTICE"),
+        );
+
+        if (matchedRules && matchedRules.length > 0) {
+          const lines = matchedRules.map((rule) => ({
+            accountingCode: rule.accountingCode,
+            isDebit: rule.isDebit,
+            particular: rule.isDebit
+              ? `支付 ${baseParsed.vendorName}`
+              : `應付 ${baseParsed.vendorName}`,
+            amount: MoneyUtil.parseInput(String(baseParsed.totalAmount || "0")),
+          }));
+
+          console.log(
+            `[VoucherLinesParsingSkill] 🎯 Stage 2 Match: Deterministic rules applied for ${baseParsed.vendorName}`,
+          );
+          return JSON.stringify({
+            generationSource: "RULE_ENGINE_STAGE_2",
+            confidence: 100,
+            aiNote: "Stage 2: Deterministic Routing Applied (TypeScript Rules)",
+            lines: lines,
+          });
+        }
+      }
     }
 
     try {
